Split README tests into numbered files with GitHub links
Replace monolithic readme_examples.rs with 10 self-contained test files
(readme_01 through readme_10), each matching a README code example exactly.
Add sections 9 (Node-Level Dynamic Range) and 10 (AmplifierModel + AM-PM)
that were previously untested. Link each README code block to its full
example on GitHub so links work from crates.io.
